fix: fix pagination controller error
